refactor: add dynamic-list-element package
