+++ packages/dynamic-list-element/package.json
@@ -1,0 +1,55 @@
+{
+  "name": "@primer/dynamic-list-element",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": {
+      "types": {
+        "import": "./dist/esm/index.d.ts",
+        "require": "./dist/cjs/index.d.cts"
+      },
+      "browser": {
+        "import": "./dist/esm/index.js",
+        "require": "./dist/cjs/index.cjs"
+      },
+      "node": {
+        "import": "./dist/esm/node/index.js",
+        "require": "./dist/cjs/node/index.cjs"
+      },
+      "import": "./dist/esm/index.js",
+      "require": "./dist/cjs/index.cjs"
+    }
+  },
+  "scripts": {
+    "clean": "rimraf dist",
+    "build": "script/build",
+    "type-check": "tsc --noEmit",
+    "watch": "rollup -c --watch"
+  },
+  "dependencies": {
+    "@lit-labs/ssr-dom-shim": "^1.3.0"
+  },
+  "devDependencies": {
+    "@rollup/plugin-inject": "^5.0.5",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "@rollup/plugin-replace": "^6.0.2",
+    "esbuild": "^0.25.3",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.30.1",
+    "rollup-plugin-esbuild": "^6.2.1",
+    "rollup-plugin-typescript2": "^0.36.0"
+  },
+  "sideEffects": [
+    "./src/define.ts",
+    "./src/index.ts",
+    "./dist/esm/define.js",
+    "./dist/esm/index.js",
+    "./dist/cjs/define.cjs",
+    "./dist/cjs/index.cjs",
+    "./dist/esm/node/define.js",
+    "./dist/esm/node/index.js",
+    "./dist/cjs/node/define.cjs",
+    "./dist/cjs/node/index.cjs"
+  ]
+}
