+++ packages/dynamic-list-element/src/dynamic-list-item-element.ts
@@ -1,0 +1,1 @@
+export class DynamicListItemElement extends HTMLElement {}
