+++ packages/dynamic-list-element/src/dynamic-list-trigger-element.ts
@@ -1,0 +1,1 @@
+export class DynamicListTriggerElement extends HTMLElement {}
