+++ packages/react/package.json
@@ -86,6 +86,7 @@
     "@lit-labs/react": "1.2.1",
     "@oddbird/popover-polyfill": "^0.4.4",
     "@primer/behaviors": "^1.8.0",
+    "@primer/dynamic-list-element": "^0.0.0",
     "@primer/live-region-element": "^0.7.1",
     "@primer/octicons-react": "^19.13.0",
     "@primer/primitives": "9.x || 10.x",
